+++ src/components/pages/home/card.type.ts
