+++ src/components/pages/home/carousel/index.tsx
