+++ src/components/pages/home/carousel/style.ts
